Update transaction pull forward for new reports
